refactored the different types of auth, cascaded changes to routes/migrations/etc
