--- api/middleware/auth.py
@@ -1,97 +1,0 @@
-from api.config import cfg
-from db.main import get_user
-from schemas.auth import TokenData, User
-
-from fastapi import (
-    Depends,
-    HTTPException,
-    status,
-)
-from fastapi.security import OAuth2PasswordBearer
-from jose import JWTError, jwt  # type: ignore
-from passlib.context import CryptContext  # type: ignore
-
-from datetime import datetime, timedelta
-from typing import Optional
-
-
-__all__ = [
-    "authenticate_user",
-    "create_access_token",
-    "get_current_active_user",
-]
-
-
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-
-def verify_password(plain_password, hashed_password):
-    return pwd_context.verify(plain_password, hashed_password)
-
-
-def get_password_hash(password):
-    return pwd_context.hash(password)
-
-
-def authenticate_user(username: str, password: str):
-    user = get_user(username)
-    if not user:
-        return False
-    if not verify_password(password, user.hashed_password):
-        return False
-    return user
-
-
-def create_access_token(
-    data: dict, expires_delta: Optional[timedelta] = None
-):
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.utcnow() + expires_delta
-    else:
-        expire = datetime.utcnow() + timedelta(minutes=15)
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(
-        to_encode,
-        cfg.auth.secret_key.get_secret_value(),
-        algorithm=cfg.auth.algorithm,
-    )
-    return encoded_jwt
-
-
-async def validate_token(
-    token: str = Depends(oauth2_scheme),
-):
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-
-    try:
-        payload = jwt.decode(
-            token,
-            cfg.auth.secret_key.get_secret_value(),
-            algorithms=[cfg.auth.algorithm],
-        )
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-        token_data = TokenData(username=username)
-
-    except JWTError:
-        raise credentials_exception
-
-    user = get_user(username=token_data.username)
-    if user is None:
-        raise credentials_exception
-    return user
-
-
-async def get_current_active_user(
-    current_user: User = Depends(validate_token),
-):
-    if current_user.disabled:
-        raise HTTPException(status_code=400, detail="Inactive user")
-    return current_user
