+++ api/routes/dependencies/__init__.py
@@ -1,2 +1,1 @@
 from api.routes.dependencies.validate_user import validate_user
-from api.routes.dependencies.current_user import active_user
