+++ api/routes/dependencies/validate_user.py
@@ -1,4 +1,4 @@
-from authentication.auth import validate_token
+from authentication.jwt_auth import validate_token
 
 
 __all__ = ["validate_user"]
