+++ api/routes/users.py
@@ -1,4 +1,5 @@
-from api.routes.dependencies import active_user
+from api.routes.dependencies import validate_user
+from db.models.users import Users
 from schemas.users import User
 
 from fastapi import Depends, APIRouter
@@ -10,8 +11,10 @@ __all__ = ["router"]
 router = APIRouter()
 
 
-@router.get("/users/me/", response_model=User)
-async def read_users_me(
-    current_user: User = Depends(active_user),
-):
-    return current_user
+@router.get("/me", response_model=User)
+async def read_users_me(username: str = Depends(validate_user)):
+    users = (
+        await Users.select().where(Users.username == username).run()
+    )
+    user = users[0]
+    return User(**user)
